SecurityBot/requirements.txt
requests
python-telegram-bot
transformers
torch
aiohttp
aiogram
requests
logging
python-telegram-bot

